Rename codelists file
This is needed to avoid problems with different functions for codelists in databuilder vs cohort-extractor
